[C] Add atomic histogram.  Clean up warnings.  Materialise imports on headers.

+++ src/test/c/hdr_histogram_log_test.c
@@ -591,5 +591,5 @@ int hdr_histogram_log_run_tests()
 
     printf("Tests run: %d\n", tests_run);
 
-    return result.message;
+    return (int) result.message;
 }
